PycharmProjects/project_ai_newsletter/requirements.txt
# Core
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0

# HTTP
httpx>=0.27.0

# RSS Parsing
feedparser>=6.0.0

# Anthropic API (for agent)
anthropic>=0.40.0

# Utilities
python-dotenv>=1.0.0

# Embeddings & Similarity
numpy>=1.24.0
scikit-learn>=1.3.0
openai>=1.0.0

# Testing
pytest>=8.0.0
